修复中转站获取保洁员姓名接口bug
